ft-lab/requirements.txt
openai>=1.40.0
pydantic>=2.6.0
python-dotenv>=1.0.1
httpx>=0.27.0
scikit-learn>=1.4.0
joblib>=1.3.0
sentence-transformers>=2.7.0
pandas>=2.2.0
tqdm>=4.66.0
